Create ebook_removal3.py
The final version of a program to take a list of all our ebooks at the library and a list of ebooks to remove from the catalog and create a list of isbn numbers so I could edit the records.
The list of removed ebooks only came with a special url used to get to the book. I had to isolate this number in the removal file and the list of all ebooks and compare them. I have tried to comment heavily so that any instruction is in the program.